from gatsby-node.js to source-plugin
Updated so the config for the query is locatead in source-plugin/gatsby-node.js.
also ran a test query on the index.js Page
