+++ source-plugin/gatsby-browser.js
@@ -1,0 +1,6 @@
+/**
+ * Implement Gatsby's Browser APIs in this file.
+ *
+ * See: https://www.gatsbyjs.org/docs/browser-apis/
+ */
+// You can delete this file if you're not using it
