+++ source-plugin/gatsby-ssr.js
@@ -3,6 +3,4 @@
  *
  * See: https://www.gatsbyjs.org/docs/ssr-apis/
  */
-
 // You can delete this file if you're not using it
-
