+++ source-plugin/index.js
@@ -1,0 +1,1 @@
+// noop
