+++ source-plugin/package.json
@@ -1,0 +1,33 @@
+{
+  "name": "gatsby-starter-plugin",
+  "version": "1.0.0",
+  "description": "A minimal boilerplate for the essential files Gatsby looks for in a plugin",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "build": "echo \"No build script setup\""
+  },
+  "keywords": [
+    "gatsby",
+    "gatsby-plugin"
+  ],
+  "author": "Kyle Gill <kyle.robert.gill@gmail.com>",
+  "license": "0BSD",
+  "dependencies": {
+    "apollo-cache-inmemory": "^1.6.6",
+    "apollo-client": "^2.6.10",
+    "apollo-link": "^1.2.14",
+    "apollo-link-http": "^1.5.17",
+    "apollo-link-rest": "^0.8.0-beta.0",
+    "apollo-link-ws": "^1.0.20",
+    "apollo-utilities": "^1.3.4",
+    "graphql": "^15.3.0",
+    "graphql-anywhere": "^4.2.7",
+    "graphql-tag": "^2.11.0",
+    "node-fetch": "^2.6.1",
+    "q": "^1.5.1",
+    "qs": "^6.9.4",
+    "subscriptions-transport-ws": "^0.9.18",
+    "ws": "^7.3.1"
+  }
+}
