+++ src/pages/index.js
@@ -1,22 +1,32 @@
 import React from "react"
-import { Link } from "gatsby"
+// import { Link } from "gatsby"
+import { graphql } from "gatsby";
 
 import Layout from "../components/layout"
-import Image from "../components/image"
-import SEO from "../components/seo"
+// import Image from "../components/image"
+// import SEO from "../components/seo"
 
-const IndexPage = () => (
-  <Layout>
-    <SEO title="Home" />
-    <h1>Hi people</h1>
-    <p>Welcome to your new Gatsby site.</p>
-    <p>Now go build something great.</p>
-    <div style={{ maxWidth: `300px`, marginBottom: `1.45rem` }}>
-      <Image />
-    </div>
-    <Link to="/page-2/">Go to page 2</Link> <br />
-    <Link to="/using-typescript/">Go to "Using TypeScript"</Link>
-  </Layout>
+
+const IndexPage = ({ data }) => (
+  <>
+    <Layout>
+      {data.allDeveloper.nodes.map(developer =>(
+          <p key={developer.id}>hi {developer.name}</p>
+      ))}
+    </Layout>
+  </>
 )
+export const query = graphql`
+  {
+    allDeveloper {
+      nodes {
+        id
+        name
+        skills
+        summary
+      }
+    }
+  }
+`
 
 export default IndexPage
